requirements.txt
attrs==20.2.0
importlib-metadata==2.0.0
iniconfig==1.0.1
lark-parser==0.10.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
six==1.15.0
toml==0.10.1
zipp==3.3.0
